Trabalhando com permissões

+++ seguranca-java-api/src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -1,34 +1,40 @@
 package br.com.forum_hub.domain.resposta;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 
+import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Service;
 
 import br.com.forum_hub.domain.topico.Status;
 import br.com.forum_hub.domain.topico.TopicoService;
 import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
+import br.com.forum_hub.services.HierarquiaService;
 import jakarta.transaction.Transactional;
 
 @Service
 public class RespostaService {
     private final RespostaRepository repository;
     private final TopicoService topicoService;
+    private final HierarquiaService hierarquiaService;
 
-    public RespostaService(RespostaRepository repository, TopicoService topicoService) {
+    public RespostaService(RespostaRepository repository, TopicoService topicoService, HierarquiaService hierarquiaService) {
         this.repository = repository;
         this.topicoService = topicoService;
+        this.hierarquiaService = hierarquiaService;
     }
 
     @Transactional
     public Resposta cadastrar(DadosCadastroResposta dados, Long idTopico, Usuario autor) {
         var topico = topicoService.buscarPeloId(idTopico);
 
-        if(!topico.estaAberto()) {
+        if (!topico.estaAberto()) {
             throw new RegraDeNegocioException("O tópico está fechado! Você não pode adicionar mais respostas.");
         }
 
-        if(topico.getQuantidadeRespostas() == 0) {
+        if (topico.getQuantidadeRespostas() == 0) {
             topico.alterarStatus(Status.RESPONDIDO);
         }
 
@@ -39,38 +45,50 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) throws AccessDeniedException {
         var resposta = buscarPeloId(dados.id());
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode atualizar essa resposta como solução!");
+
         return resposta.atualizarInformacoes(dados);
     }
 
-    public List<Resposta> buscarRespostasTopico(Long id){
+    public List<Resposta> buscarRespostasTopico(Long id) {
         return repository.findByTopicoId(id);
     }
 
     @Transactional
-    public Resposta marcarComoSolucao(Long id) {
+    public Resposta marcarComoSolucao(Long id, Usuario logado) throws AccessDeniedException {
         var resposta = buscarPeloId(id);
 
         var topico = resposta.getTopico();
-        if(topico.getStatus() == Status.RESOLVIDO)
-            throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_INSTRUTOR"))
+            throw new AccessDeniedException("Você não pode marcar essa resposta como solução!");
+
+        if (topico.getStatus() == Status.RESOLVIDO)
+            throw new RegraDeNegocioException(
+                    "O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
 
         topico.alterarStatus(Status.RESOLVIDO);
         return resposta.marcarComoSolucao();
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) throws AccessDeniedException {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode Deletar essa resposta como solução!");
 
         repository.deleteById(id);
 
         topico.decrementarRespostas();
         if (topico.getQuantidadeRespostas() == 0)
             topico.alterarStatus(Status.NAO_RESPONDIDO);
-        else if(resposta.ehSolucao())
+        else if (resposta.ehSolucao())
             topico.alterarStatus(Status.RESPONDIDO);
     }
